feat: 언어별 색상 매핑 함수 추가 및 관련 코드 정리
