+++ src/pages/RepositoryPage.jsx
@@ -30,6 +30,7 @@ import {
 } from 'lucide-react';
 import DashboardLayout from '../components/dashboard-layout';
 import { getRepositoryDetails } from '../services/repositoryService';
+import { getLanguageColor } from '../utils/languageUtils';
 
 export default function RepositoryPage() {
   const { id: repoId } = useParams();
@@ -180,33 +181,6 @@ export default function RepositoryPage() {
   const formatDate = (dateString) => {
     if (!dateString) return '정보 없음';
     return new Date(dateString).toLocaleDateString('ko-KR');
-  };
-
-  // 언어별 색상 매핑 함수
-  const getLanguageColor = (language) => {
-    const colors = {
-      JavaScript: '#f1e05a',
-      TypeScript: '#2b7489',
-      Python: '#3572A5',
-      Java: '#b07219',
-      'C++': '#f34b7d',
-      C: '#555555',
-      Go: '#00ADD8',
-      Rust: '#dea584',
-      PHP: '#4F5D95',
-      Ruby: '#701516',
-      Swift: '#ffac45',
-      Kotlin: '#F18E33',
-      HTML: '#e34c26',
-      CSS: '#1572B6',
-      Shell: '#89e051',
-      Assembly: '#6E4C13',
-      'Jupyter Notebook': '#DA5B0B',
-      XML: '#0060ac',
-      Gradle: '#02303a',
-      Perl: '#0298c3',
-    };
-    return colors[language] || '#8b5cf6';
   };
 
   return (
